Refactor and fix bug with actions
