+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"C:\\Desarrollo\\Repository\\React Repository\\bookstore\\src\\index.js":"1","C:\\Desarrollo\\Repository\\React Repository\\bookstore\\src\\component\\App.js":"2","C:\\Desarrollo\\Repository\\React Repository\\bookstore\\src\\reducers\\index.js":"3","C:\\Desarrollo\\Repository\\React Repository\\bookstore\\src\\component\\BookList.js":"4","C:\\Desarrollo\\Repository\\React Repository\\bookstore\\src\\component\\BookForm.js":"5","C:\\Desarrollo\\Repository\\React Repository\\bookstore\\src\\reducers\\book.js":"6","C:\\Desarrollo\\Repository\\React Repository\\bookstore\\src\\component\\Book.js":"7","C:\\Desarrollo\\Repository\\React Repository\\bookstore\\src\\actions\\index.js":"8"},{"size":694,"mtime":1612298317521,"results":"9","hashOfConfig":"10"},{"size":254,"mtime":1612287217821,"results":"11","hashOfConfig":"10"},{"size":169,"mtime":1612287699398,"results":"12","hashOfConfig":"10"},{"size":762,"mtime":1612298317521,"results":"13","hashOfConfig":"10"},{"size":1493,"mtime":1612304383358,"results":"14","hashOfConfig":"10"},{"size":292,"mtime":1612303513875,"results":"15","hashOfConfig":"10"},{"size":407,"mtime":1612298317521,"results":"16","hashOfConfig":"10"},{"size":0,"mtime":1612212731446,"results":"17","hashOfConfig":"10"},{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1u47te5",{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"26","messages":"27","errorCount":1,"warningCount":3,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"28","messages":"29","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"30","messages":"31","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"32","messages":"33","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\Desarrollo\\Repository\\React Repository\\bookstore\\src\\index.js",[],"C:\\Desarrollo\\Repository\\React Repository\\bookstore\\src\\component\\App.js",[],"C:\\Desarrollo\\Repository\\React Repository\\bookstore\\src\\reducers\\index.js",[],"C:\\Desarrollo\\Repository\\React Repository\\bookstore\\src\\component\\BookList.js",[],"C:\\Desarrollo\\Repository\\React Repository\\bookstore\\src\\component\\BookForm.js",["34","35","36","37"],"C:\\Desarrollo\\Repository\\React Repository\\bookstore\\src\\reducers\\book.js",[],"C:\\Desarrollo\\Repository\\React Repository\\bookstore\\src\\component\\Book.js",[],"C:\\Desarrollo\\Repository\\React Repository\\bookstore\\src\\actions\\index.js",[],{"ruleId":"38","severity":1,"message":"39","line":13,"column":5,"nodeType":"40","messageId":"41","endLine":13,"endColumn":16},{"ruleId":"38","severity":1,"message":"39","line":16,"column":9,"nodeType":"40","messageId":"41","endLine":16,"endColumn":20},{"ruleId":"38","severity":1,"message":"39","line":37,"column":5,"nodeType":"40","messageId":"41","endLine":37,"endColumn":16},{"ruleId":"42","severity":2,"message":"43","line":38,"column":11,"nodeType":"44","endLine":38,"endColumn":21},"no-console","Unexpected console statement.","MemberExpression","unexpected","react/prop-types","'createBook' is missing in props validation","Identifier"]
+[{"/home/paul/Documents/code/MV/technical projects/react/bookstore/src/index.js":"1","/home/paul/Documents/code/MV/technical projects/react/bookstore/src/component/App.js":"2","/home/paul/Documents/code/MV/technical projects/react/bookstore/src/reducers/index.js":"3","/home/paul/Documents/code/MV/technical projects/react/bookstore/src/reducers/book.js":"4","/home/paul/Documents/code/MV/technical projects/react/bookstore/src/component/Book.js":"5","/home/paul/Documents/code/MV/technical projects/react/bookstore/src/actions/index.js":"6","/home/paul/Documents/code/MV/technical projects/react/bookstore/src/containers/BookForm.js":"7","/home/paul/Documents/code/MV/technical projects/react/bookstore/src/containers/BookList.js":"8","/home/paul/Documents/code/MV/technical projects/react/bookstore/src/actions/booksActions.js":"9"},{"size":694,"mtime":1612345438509,"results":"10","hashOfConfig":"11"},{"size":278,"mtime":1612369196241,"results":"12","hashOfConfig":"11"},{"size":169,"mtime":1612345438513,"results":"13","hashOfConfig":"11"},{"size":417,"mtime":1612373057528,"results":"14","hashOfConfig":"11"},{"size":407,"mtime":1612345438477,"results":"15","hashOfConfig":"11"},{"size":93,"mtime":1612373595236,"results":"16","hashOfConfig":"11"},{"size":1830,"mtime":1612373742507,"results":"17","hashOfConfig":"11"},{"size":773,"mtime":1612369196237,"results":"18","hashOfConfig":"11"},{"size":188,"mtime":1612373564208,"results":"19","hashOfConfig":"11"},{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},"5o415q",{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},{"filePath":"27","messages":"28","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},{"filePath":"31","messages":"32","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},{"filePath":"33","messages":"34","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"35","messages":"36","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},{"filePath":"37","messages":"38","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/home/paul/Documents/code/MV/technical projects/react/bookstore/src/index.js",[],["39","40","41","42","43","44"],"/home/paul/Documents/code/MV/technical projects/react/bookstore/src/component/App.js",[],"/home/paul/Documents/code/MV/technical projects/react/bookstore/src/reducers/index.js",[],"/home/paul/Documents/code/MV/technical projects/react/bookstore/src/reducers/book.js",[],"/home/paul/Documents/code/MV/technical projects/react/bookstore/src/component/Book.js",[],"/home/paul/Documents/code/MV/technical projects/react/bookstore/src/actions/index.js",[],"/home/paul/Documents/code/MV/technical projects/react/bookstore/src/containers/BookForm.js",[],"/home/paul/Documents/code/MV/technical projects/react/bookstore/src/containers/BookList.js",[],"/home/paul/Documents/code/MV/technical projects/react/bookstore/src/actions/booksActions.js",[],{"ruleId":"45","replacedBy":"46"},{"ruleId":"47","replacedBy":"48"},{"ruleId":"49","replacedBy":"50"},{"ruleId":"51","replacedBy":"52"},{"ruleId":"53","replacedBy":"54"},{"ruleId":"55","replacedBy":"56"},"lines-around-directive",["57"],"no-spaced-func",["58"],"global-require",[],"no-buffer-constructor",[],"no-new-require",[],"no-path-concat",[],"padding-line-between-statements","func-call-spacing"]
