+++ src/actions/index.js
@@ -1,0 +1,3 @@
+import { createBook, removeBook } from './booksActions';
+
+export { createBook, removeBook };
