+++ src/reducers/book.js
@@ -1,8 +1,13 @@
 const BookReducer = (state = [], action) => {
   switch (action.type) {
     case 'CREATE_BOOK': return {
-      ...state,
-
+      books: [
+        {
+          id: Math.random(),
+          title: 'Testing',
+          category: 'Category Testing',
+        },
+      ],
     };
     case 'REMOVE_BOOK': return {
       ...state,
